+++ pkg/objectcache/containerprofilecache/containerprofilecache.go
@@ -613,62 +613,5 @@ func (c *ContainerProfileCacheImpl) waitForSharedContainerData(containerID strin
 	}, backoff.WithBackOff(backoff.NewExponentialBackOff()))
 }
 
-// ReconcileOnce is an exported thin wrapper around reconcileOnce for use by
-// out-of-package integration tests (e.g. tests/containerprofilecache/).
-// Production code should use tickLoop / Start.
-func (c *ContainerProfileCacheImpl) ReconcileOnce(ctx context.Context) {
-	c.reconcileOnce(ctx)
-}
-
-// SeedEntryForTest directly inserts a CachedContainerProfile entry keyed by
-// containerID. Intended exclusively for out-of-package integration tests that
-// cannot call the internal addContainer path. Do not call from production code.
-func (c *ContainerProfileCacheImpl) SeedEntryForTest(containerID string, entry *CachedContainerProfile) {
-	c.entries.Set(containerID, entry)
-}
-
-// RefreshAllEntriesForTest is an exported thin wrapper around refreshAllEntries
-// for use by out-of-package integration tests. Do not call from production code.
-func (c *ContainerProfileCacheImpl) RefreshAllEntriesForTest(ctx context.Context) {
-	c.refreshAllEntries(ctx)
-}
-
-// WarmContainerLocksForTest acquires and immediately releases each container
-// lock, initialising the internal SafeMap so concurrent callers don't trigger
-// the goradd/maps nil-check-before-lock initialisation race (pre-existing
-// upstream bug in SafeMap v1.3.0). Do not call from production code.
-func (c *ContainerProfileCacheImpl) WarmContainerLocksForTest(ids []string) {
-	for _, id := range ids {
-		c.containerLocks.WithLock(id, func() {})
-	}
-}
-
-// WarmPendingForTest initialises the internal pending SafeMap by writing then
-// deleting a nil entry for each id, ensuring the map is non-nil before the
-// concurrent phase. Prevents the goradd/maps nil-check-before-lock
-// initialisation race in SafeMap.Len / SafeMap.Delete. Do not call from
-// production code.
-func (c *ContainerProfileCacheImpl) WarmPendingForTest(ids []string) {
-	for _, id := range ids {
-		c.pending.Set(id, nil)
-		c.pending.Delete(id)
-	}
-}
-
-// SeedEntryWithOverlayForTest seeds a CachedContainerProfile entry with user
-// AP and NN overlay refs set. Used by out-of-package integration tests that
-// need to exercise the refresh overlay path but cannot construct the internal
-// namespacedName type directly. Pass empty strings to leave refs nil.
-// Do not call from production code.
-func (c *ContainerProfileCacheImpl) SeedEntryWithOverlayForTest(containerID string, entry *CachedContainerProfile, apNS, apName, nnNS, nnName string) {
-	if apName != "" {
-		entry.UserAPRef = &namespacedName{Namespace: apNS, Name: apName}
-	}
-	if nnName != "" {
-		entry.UserNNRef = &namespacedName{Namespace: nnNS, Name: nnName}
-	}
-	c.entries.Set(containerID, entry)
-}
-
 // Ensure ContainerProfileCacheImpl implements the ContainerProfileCache interface.
 var _ objectcache.ContainerProfileCache = (*ContainerProfileCacheImpl)(nil)
